Move implementation from hpp to cpp
